Database_Implementation 1.04
Found the issue may be due to how Javascript's asynchronous processing of data, the code to push the data into the necessary var's are being populated after rendering, hence the late updating of renders and only updating after the second time of pressing the button, calling the api. Implemented a possible fix via asynchronous calls but implementation may be improper as I created this while in class, will look over and confirm/settle bugs later.