DB_Implementation_VERY_ROUGH_FINAL
Drew's Final push -

Added a login kind of flow to the project. However, the children are not yet excusive to a particular account because I could not get the "Parent_ID" to be pulled from the db no matter what I do. Logins work though via what data is in the db for the users we assign as parents.